update layers dao and services

+++ src/app/entities/exchange_rate.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass(frozen=True)
@@ -6,4 +7,4 @@ class ExchangeRate:
     id: int
     base_currency_id: int
     target_currency_id: int
-    rate: float
+    rate: Decimal
